fix(build): move useKeyboardNavigation so its import resolves
PR #714 committed the hook to the repo root while App.tsx imports it as a
sibling of itself. tsconfig only includes src, so the file was never even
type-checked and the import never resolved — main fails to compile with
TS2307 before any other change is applied.

Relocating the hook alongside the other app-wide hooks fixes it. No lines
of the hook itself changed.

This does not touch the hook's runtime behaviour. It binds keydown on
window in capture phase and calls stopPropagation on Escape, the arrows,
w/a/s/d and Space, which silently kills every window-bubble keydown
listener in the app — nearly a hundred of them, including the player's.
Left for its author.
